Final tests passing

+++ c_messageparser_tests.h
@@ -3,6 +3,6 @@
 #include <unistd.h>
 
 typedef struct thread_args {
-    int destination_id;
+    unsigned long int destination_id;
 } thread_args;
 sem_t test_start;
